Separação dos menus, main e Homepage
