+++ lib/MenuAluno.dart
@@ -3,4 +3,11 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'main.dart';
 
+class MenuAluno extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Placeholder();
+  }
+}
+
 //nao mergear com develop
