+++ lib/bak.txt
@@ -1,0 +1,46 @@
+    /*var selectedIndex = 0;
+    
+      switch (selectedIndex) {
+      case 0:
+        page = Placeholder();
+        break;
+      case 1:
+        page = Placeholder();
+        break;
+      case 2:
+        page = Placeholder();
+        break;
+      case 3:
+        page = Placeholder();
+        break;
+      default:
+        throw UnimplementedError('no widget for $selectedIndex');
+    }*/
+
+    NavigationRail(
+                extended: constraints.maxWidth >= 600, // ← Here.
+                destinations: [
+                  NavigationRailDestination(
+                    icon: Icon(Icons.menu),
+                    label: Text('Menu'),
+                  ),
+                  NavigationRailDestination(
+                    icon: Icon(Icons.folder),
+                    label: Text('Cursos'),
+                  ),
+                  NavigationRailDestination(
+                    icon: Icon(Icons.menu_book),
+                    label: Text('Testes'),
+                  ),
+                  NavigationRailDestination(
+                    icon: Icon(Icons.logout),
+                    label: Text('Log out'),
+                  ),
+                ],
+                selectedIndex: selectedIndex,
+                onDestinationSelected: (value) {
+                  setState(() {
+                    selectedIndex = value;
+                  });
+                },
+              ),
