+++ lib/main.dart
@@ -1,12 +1,7 @@
-import 'package:english_words/english_words.dart';
 import 'package:flutter/material.dart';
-import 'package:projeto_final_front/Cadastro.dart';
+import 'package:projeto_final_front/MenuAluno.dart';
 import 'package:provider/provider.dart';
-import 'MenuAluno.dart';
-import 'MenuCursos.dart';
-import 'MenuTestes.dart';
-import 'Login.dart';
-import 'Cadastro.dart';
+import 'HomePage.dart';
 
 void main() {
   runApp(MyApp());
@@ -37,78 +32,7 @@ class MyAppState extends ChangeNotifier {
   void addTest(String teste) {
     test.add(teste);
   }
-}
 
-class MyHomePage extends StatefulWidget {
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  var selectedIndex = 0;
-
-  @override
-  Widget build(BuildContext context) {
-    Widget page;
-    switch (selectedIndex) {
-      case 0:
-        page = LoginPage();
-        break;
-      case 1:
-        page = CursosPage();
-        break;
-      case 2:
-        page = CadastroPage();
-        break;
-      case 3:
-        page = Placeholder();
-        break;
-      default:
-        throw UnimplementedError('no widget for $selectedIndex');
-    }
-
-    return LayoutBuilder(builder: (context, constraints) {
-      return Scaffold(
-        body: Row(
-          children: [
-            SafeArea(
-              child: NavigationRail(
-                extended: constraints.maxWidth >= 600, // ← Here.
-                destinations: [
-                  NavigationRailDestination(
-                    icon: Icon(Icons.menu),
-                    label: Text('Menu'),
-                  ),
-                  NavigationRailDestination(
-                    icon: Icon(Icons.folder),
-                    label: Text('Cursos'),
-                  ),
-                  NavigationRailDestination(
-                    icon: Icon(Icons.menu_book),
-                    label: Text('Testes'),
-                  ),
-                  NavigationRailDestination(
-                    icon: Icon(Icons.logout),
-                    label: Text('Log out'),
-                  ),
-                ],
-                selectedIndex: selectedIndex,
-                onDestinationSelected: (value) {
-                  setState(() {
-                    selectedIndex = value;
-                  });
-                },
-              ),
-            ),
-            Expanded(
-              child: Container(
-                color: Theme.of(context).colorScheme.primaryContainer,
-                child: page,
-              ),
-            ),
-          ],
-        ),
-      );
-    });
-  }
+  var logado = false;
+  var tipo_logado = 0;
 }
